Add files via upload
n this example, we create a parent div with a class of "parent", and four child divs with a class of "child". We use display: flex on the parent to enable flexbox, and justify-content: space-between to distribute the child elements evenly with space between them. We also use align-items: center to vertically center the child elements within the parent.

Each child div has a width of 100px, a height of 80px, and a gray background color. We use display: flex, justify-content: center, and align-items: center to center the text within each child element.